+++ NEWS.md
@@ -8,14 +8,15 @@
   
 ## New features
 
-- Add vignette "Reading and Displaying inForm Image Files"
-- Add vignette "Computing Inter-cellular Distances"
 - `spatial_distribution_report` creates an HTML report showing the 
   location and nearest-neighbor relations of cells in a single field.
 - `read_components` reads component image files.
 - `count_within` counts the number of `from` cells having a `to` cell
    within a given radius.
 - `list_cell_seg_files` lists all cell seg data files in a folder.
+- Added vignettes
+    - "Reading and Displaying inForm Image Files"
+    - "Computing Inter-cellular Distances"
 
 ## Bug fixes